pyproject.toml
[project]
name = "vitex"
version = "0.1.0"
description = "Vitex Internal RAG Chatbot"
package-mode = false
authors = [
    {name = "Darwin"}
]
requires-python = ">=3.11, <3.14"

dependencies = [
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-huggingface>=0.1.0",
    "faiss-cpu>=1.8.0",
    "sentence-transformers>=3.0.0",
    "pypdf>=5.0.0",
    "docx2txt>=0.8",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "python-dotenv>=1.0.0",
    "langchainhub>=0.1.20",
    "pydantic>=2.0.0",
    "httplib2 (>=0.31.2,<0.32.0)",
    "langchain-classic (>=1.0.3,<2.0.0)",
    "rank-bm25 (>=0.2.2,<0.3.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"